docs: further improvements
