+++ mkdocs.yml
@@ -42,81 +42,22 @@ markdown_extensions:
   - pymdownx.superfences
 
 nav:
-  - Overview: README.md
-  - guides/quick-start.md
-  - Guides:
+  - installation.md
+  - quick-start.md
+  - Usage Guides:
+    - guides/contracts.md
     - guides/contract-events.md
     - guides/oracle-usage.md
     - guides/aens-usage.md
-    - guides/contract-aci-usage.md
     - guides/delegate-signature-to-contract.md
-    - Wallet Api: guides/how-to-build-aepp-using-new-wallet-api.md
-    - Building Wallets: guides/how-to-build-an-wallet-app-or-extension.md
+    - Wallet Api: guides/how-to-build-aepp.md
+    - Building Wallets: guides/how-to-build-wallet.md
     - guides/low-vs-high-usage.md
     - guides/import-tree-shaking.md
-    - Migration:
-      - guides/migration/migration-7.0.0.md
-  - API Reference:
-    - account:
-      - api/account.md
-      - api/account/selector.md
-      - api/account/memory.md
-      - Account Handling: api/accounts.md
-    - ae:
-      - api/ae/aens.md
-      - api/ae/aepp.md
-      - api/ae/contract.md
-      - api/ae/oracle.md
-      - api/ae/universal.md
-      - api/ae/wallet.md
-    - chain:
-      - api/chain.md
-      - api/chain/node.md
-    - channel: api/channel/index.md
-    - contract:
-      - api/contract/aci.md
-      - api/contract/compiler.md
-      - Generalized Accounts: api/contract/ga.md
-    - oracle:
-      - api/oracle.md
-      - api/oracle/node.md
-    - Deprecated RPC Client: api/rpc/client.md
-    - tx:
-      - builder:
-        - api/tx/builder.md
-        - api/tx/builder/helpers.md
-        - api/tx/builder/schema.md
-      - tx: api/tx.md
-      - tx stamp: api/tx/tx.md
-      - tx object: api/tx/tx-object.md
-      - tx validator: api/tx/validator.md
-    - Aepp Wallet Communication:
-      - Connection:
-        - Setup: api/utils/aepp-wallet-communication/connection.md
-        - api/utils/aepp-wallet-communication/content-script-bridge.md
-        - api/utils/aepp-wallet-communication/wallet-detector.md
-        - api/utils/aepp-wallet-communication/connection/browser-runtime.md
-        - Window message: api/utils/aepp-wallet-communication/connection/browser-window-message.md
-      - RPC:
-        - Aepp RPC: api/utils/aepp-wallet-communication/rpc/aepp-rpc.md
-        - RPC client: api/utils/aepp-wallet-communication/rpc/rpc-client.md
-        - Wallet RPC: api/utils/aepp-wallet-communication/rpc/wallet-rpc.md
-    - utils:
-      - AE Utilities: api/ae.md
-      - api/utils/amount-formatter.md
-      - api/utils/bignumber.md
-      - api/utils/bytes.md
-      - api/utils/crypto.md
-      - api/utils/keystore.md
-      - api/contract.md
-      - api/node.md
-      - api/node-pool.md
-      - api/utils/swagger.md
-  - Examples:
+  - Tutorials:
     - NodeJS:
-      - Setup: guides/import-nodejs.md
-      - examples/node/wallet.md
-      - examples/node/contract.md
-    - JS:
-      - guides/import-script-tag.md
-      - guides/import-vuejs.md
+      - examples/node/transfer-ae-tokens.md
+      - examples/node/contract-interaction.md
+    - tutorials/vuejs/helloworld-blockheight.md
+  - Migration Guides:
+    - guides/migration/migration-7.0.0.md
