+++ package.json
@@ -96,7 +96,7 @@
     "standard-version": "^9.3.0",
     "ts-node": "^10.0.0",
     "ts-standard": "^10.0.0",
-    "typescript": "^4.3.2",
+    "typescript": "4.2.4",
     "webpack": "^5.38.1",
     "webpack-bundle-analyzer": "^4.4.2",
     "webpack-cli": "4.7.0"
