+++ tooling/docs/header.hbs
@@ -1,3 +1,3 @@
 <a id="{{{anchorName}}}"></a>
 
-{{!-- {{>heading-indent}}{{>sig-name}} --}}
+{{>heading-indent}}{{>sig-name}}
